Upgrade Error / Alias Handling.
Instead of automatically redirecting to an add page, you will go to an error page, which asks you to create a custom error page first.  Once you create a custom error page, then when you visit a url which doesn't exist you will get prompted to add the page if you want.  It will not auto go to the add method, unless you click the add link on the custom error page.